Fix link to new YouTube channel
Catched up with your live stream the last days and wanted to subscribe to your new YouTube channel by clicking the link in the show notes. While it was indeed a YouTube link it took me to the SICP YouTube playlist like the URL  two lines above. I took the link from your notes from the live stream on 2021-11-26 and replaced it. Thanks for the great content, btw.!